+++ session/sessiontest/BUILD.bazel
@@ -26,6 +26,7 @@ go_test(
         "//privilege/privileges",
         "//session",
         "//sessionctx",
+        "//sessionctx/stmtctx",
         "//sessionctx/variable",
         "//store/copr",
         "//store/mockstore",
