+++ sessionctx/stmtctx/stmtctx.go
@@ -464,7 +464,6 @@ func (sh *StmtHints) Clone() *StmtHints {
 	return &StmtHints{
 		MemQuotaQuery:                  sh.MemQuotaQuery,
 		MaxExecutionTime:               sh.MaxExecutionTime,
-		TidbKvReadTimeout:              sh.TidbKvReadTimeout,
 		ReplicaRead:                    sh.ReplicaRead,
 		AllowInSubqToJoinAndAgg:        sh.AllowInSubqToJoinAndAgg,
 		NoIndexMergeHint:               sh.NoIndexMergeHint,
@@ -476,7 +475,6 @@ func (sh *StmtHints) Clone() *StmtHints {
 		HasMemQuotaHint:                sh.HasMemQuotaHint,
 		HasReplicaReadHint:             sh.HasReplicaReadHint,
 		HasMaxExecutionTime:            sh.HasMaxExecutionTime,
-		HasTidbKvReadTimeout:           sh.HasTidbKvReadTimeout,
 		HasEnableCascadesPlannerHint:   sh.HasEnableCascadesPlannerHint,
 		HasResourceGroup:               sh.HasResourceGroup,
 		SetVars:                        vars,
